+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `CadenzaClient#fetchObjectInfo`
 
+### Fixed
+- Make Cadenza return an error instead of showing an error page.
+
+
 ## 2.10.0 - 2024-06-11
 ### Added
 - `CadenzaClient#destroy`